extract PaginatedQualifiedConversationIDList
